LPD-19822 Pull out Criteria/Builder as pojo class
