+++ modules/dxp/apps/portal-search-tuning/portal-search-tuning-rankings-web/src/main/java/com/liferay/portal/search/tuning/rankings/web/internal/index/DuplicateQueryStringsDetector.java
@@ -5,9 +5,6 @@
 
 package com.liferay.portal.search.tuning.rankings.web.internal.index;
 
-import com.liferay.portal.search.tuning.rankings.index.name.RankingIndexName;
-
-import java.util.Collection;
 import java.util.List;
 
 /**
@@ -15,45 +12,6 @@ import java.util.List;
  */
 public interface DuplicateQueryStringsDetector {
 
-	public Criteria.Builder builder();
-
 	public List<String> detect(Criteria criteria);
-
-	public interface Criteria {
-
-		public String getGroupExternalReferenceCode();
-
-		public String getIndex();
-
-		public Collection<String> getQueryStrings();
-
-		public RankingIndexName getRankingIndexName();
-
-		public String getSXPBlueprintExternalReferenceCode();
-
-		public String getUnlessRankingDocumentId();
-
-		public interface Builder {
-
-			public Criteria build();
-
-			public Builder groupExternalReferenceCode(
-				String groupExternalReferenceCode);
-
-			public Builder index(String index);
-
-			public Builder queryStrings(Collection<String> queryStrings);
-
-			public Builder rankingIndexName(RankingIndexName rankingIndexName);
-
-			public Builder sxpBlueprintExternalReferenceCode(
-				String sxpBlueprintExternalReferenceCode);
-
-			public Builder unlessRankingDocumentId(
-				String unlessRankingDocumentId);
-
-		}
-
-	}
 
 }
